requirements.txt
aiogram==3.22.0
aiohttp==3.12.15
asyncpg==0.30.0
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
typing_extensions==4.15.0
